Moved code from __main__ into main and added helper script jenkinsfilegenerator in pyproject.toml
